Don't copy over input to output
This is a hangover from Pandoc. Pandoc wasn't smart enough to
copy over the files it needed, so we needed to completely dupe
the input directory to the output directory.

LaTeXML is much smarter, so this is no longer needed, and we're
just wasting time and storage space doing it.

Because we're no longer copying stuff, this is an opportunity to
clean things up, and clearly separate input and output code.