Render gzipped tex files that aren't in a tarball
This is a bit of a hack to allow us to support Arxiv's bulk source
files. They provide lots of .gz files that could either be a
tarball or a gzipped TeX file, but they don't tell us what it is.

So, this extracts .gz files as a tarball, and if that fails, assume
it is a tex file.

As part of this we switched from the "tar" NPM package to just
shelling out to tar. This is because the NPM package doesn't
throw an error if the file isn't a tar file. Sigh.
https://github.com/npm/node-tar/issues/137

Closes #276