Don't mount node_modules as volume in development
It takes ~5s to mount, which is infuriating when developing.
This hugely speeds up all the development commands. Tests seem to
run a lot faster than 5s for some reason (44s instead of >1m).

The only downside of this approach is that node_modules in the
local dir is going to fuck things up. We should make a check
for that.